chore: clean main
